+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 <!-- Rename `--busy`/`__busy` CSS classes to `--wait`/`__wait` -->
 
+<!-- (breaking change) Add `background-color: transparent;` to `TextInput` -->
+
 0.14.25 / 22.08.2018
 ====================